Separate out subscription access implementations; add method for getting subscriber addresses
